PARQUET-369: Add shaded SLF4J NOP binding.
This silences the complaint that no logger implementation could be
found. No logger implementation was possible because the class that
SLF4J was trying to load had been relocated. The only options are to
relocate an implementation along with SLF4J or not shade SLF4J. This
adds the NOP logger to silence the warning.